+++ test.js
@@ -1,0 +1,17 @@
+'use strict'
+
+/**
+ * Dependencies
+ */
+
+const assert = require('assert')
+
+/**
+ * Assertions
+ */
+
+describe("index.js", () => {
+  it("should have tests", () => {
+    assert.ok(true)
+  })
+})